Remove -l flag from lpr command
The -l flag causes other flags (e.g. sides and number of copies) not to have effect. Removing it makes the other flags work but slows down printing for very complicated pdfs (the job takes a while to show up in the print queue and the printer sometimes hangs between pages).